Implement Pipeline to combine Loader and Parser
