+++ reader.go
@@ -2,8 +2,6 @@ package rlogs
 
 import (
 	"fmt"
-
-	"github.com/pkg/errors"
 )
 
 // Reader is interface of log load and parse
@@ -19,9 +17,8 @@ type BasicReader struct {
 
 // LogEntry is a pair of Loader and Paresr
 type LogEntry struct {
-	Src LogSource
-	Ldr Loader
-	Psr Parser
+	Src  LogSource
+	Pipe Pipeline
 }
 
 func (x *BasicReader) Read(src LogSource) chan *LogQueue {
@@ -29,42 +26,22 @@ func (x *BasicReader) Read(src LogSource) chan *LogQueue {
 	if x.QueueSize > 0 {
 		queueSize = x.QueueSize
 	}
-
 	ch := make(chan *LogQueue, queueSize)
-	go func() {
-		defer close(ch)
 
-		var entry *LogEntry
-		for _, e := range x.LogEntries {
-			if e.Src.Contains(src) {
-				entry = e
-				break
-			}
+	var entry *LogEntry
+	for _, e := range x.LogEntries {
+		if e.Src.Contains(src) {
+			entry = e
+			break
 		}
+	}
 
-		if entry == nil {
-			ch <- &LogQueue{Error: fmt.Errorf("No matched LogEntry for %v", src)}
-			return
-		}
+	if entry == nil {
+		ch <- &LogQueue{Error: fmt.Errorf("No matched LogEntry for %v", src)}
+		return nil
+	}
 
-		msgch := entry.Ldr.Load(src)
-		for msg := range msgch {
-			if msg.Error != nil {
-				ch <- &LogQueue{Error: errors.Wrap(msg.Error, "Fail to load log message")}
-				return
-			}
-
-			logs, err := entry.Psr.Parse(msg)
-			if err != nil {
-				ch <- &LogQueue{Error: errors.Wrap(msg.Error, "Fail to parse log message")}
-				return
-			}
-
-			for i := range logs {
-				ch <- &LogQueue{Log: logs[i]}
-			}
-		}
-	}()
+	go entry.Pipe.Run(src, ch)
 
 	return ch
 }
